requirements.dev.txt
Django==3.2.10
django-environ==0.4
psycopg2-binary==2.9.3
django-cors-headers==3.11.0
gunicorn==20.1.0
django-celery-results==2.4.0
python-gettext==4.0
django-filter==22.1
six==1.16.0
celery[redis]==5.2.7
redis==4.3.4
Pillow==9.2.0
requests==2.28.1
daphne==4.0.0
channels-redis==4.0.0
django-crispy-forms==2.0